backend/requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
requests>=2.31.0
google-genai>=0.3.0
numpy>=1.26.0
python-dotenv>=1.0.0